pyproject.toml
[project]
name = "cosecha-urbana-ai"
version = "0.1.0"
description = "AI agent for food waste redistribution from malls to shelters"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    {name = "Said Moreno", email = "said@cosechaurbana.ai"}
]

dependencies = [
    "fastapi>=0.111",
    "uvicorn[standard]>=0.29",
    "pydantic>=2.7",
    "pydantic-settings>=2.3",
    "email-validator>=2.1",
    "elasticsearch[async]>=8.13",
    "langchain>=0.2",
    "langchain-groq>=0.2",
    "langchain-anthropic>=0.1",
    "langchain-elasticsearch>=0.2",
    "langgraph>=0.1",
    "anthropic>=0.28",
    "httpx>=0.27",
    "structlog>=24.1",
    "tenacity>=8.3",
    "python-dotenv>=1.0",
    "geopy>=2.4",
    "anyio>=4.3",
    "typing-extensions>=4.11",
]

[project.urls]
Repository = "https://github.com/saidmoreno808/cosecha_urbana_ai"

[tool.uv]
dev-dependencies = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4.0",
    "httpx>=0.27",
    "faker>=24.0",
    "ruff>=0.4",
    "mypy>=1.9",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = false
ignore_missing_imports = true
